Cleanup: Move CLI related files to a subpackage.
